src/snyk_collections_by_path.py
import snyk

def add_projects_to_collections(target, path_to_collection_map):
    # Initialize a Snyk client with your token
    client = snyk.SnykClient("<your_snyk_token>")

    # Get the organization
    org = client.organizations.get("<your_organization_id>")

    # Get the target
    target = org.projects.get(target)

    # Loop through the projects
    for project in target.projects.all():
        # Get the project's file path
        file_path = project.file_path

        # Loop through the starting paths
        for start_path, collection in path_to_collection_map.items():
            # If the file path starts with the starting path, add the project to the collection
            if file_path.startswith(start_path):
                collection.append(project)

if __name__ == "__main__":
    # Define the starting paths and the collections
    path_to_collection_map = {
        "path1": [],
        "path2": [],
        # Add more paths and collections as needed
    }

    # Call the function
    add_projects_to_collections("<your_target>", path_to_collection_map)

    # Print the collections
    for path, collection in path_to_collection_map.items():
        print(f"Projects in {path}:")
        for project in collection:
            print(project.name)